+++ Tests/Unit/Stubs/Generic.psm1
@@ -1690,3 +1690,169 @@ function New-AddressList
         $Confirm
     )
 }
+
+function Connect-Graph
+{
+    [CmdletBinding()]
+    Param(
+        [Parameter()]
+        [System.String[]]
+        $Scopes
+    )
+}
+
+function Get-MGGroupPlannerPlan
+{
+    [CmdletBinding()]
+    Param(
+        [Parameter()]
+        [System.String]
+        $GroupId
+    )
+}
+
+function Update-MGPlannerPlan
+{
+    [CmdletBinding()]
+    Param(
+        [Parameter()]
+        [System.String]
+        $Owner,
+
+        [Parameter()]
+        [System.String]
+        $Title,
+
+        [Parameter()]
+        [System.String]
+        $PlannerPlanId
+    )
+}
+
+function Get-MGPlannerTask
+{
+    [CmdletBinding()]
+    Param(
+        [Parameter()]
+        [System.String]
+        $PlannerTaskId
+    )
+}
+
+function New-MGPlannerTask
+{
+    [CmdletBinding()]
+    Param(
+        [Parameter()]
+        [System.String]
+        $PlanId,
+
+        [Parameter()]
+        [System.String]
+        $Title,
+
+        [Parameter()]
+        [System.String]
+        $Description,
+
+        [Parameter()]
+        [System.String]
+        $StartDateTime,
+
+        [Parameter()]
+        [System.String]
+        $CompletedDateTime,
+
+        [Parameter()]
+        [ValidateRange(0, 100)]
+        [System.Uint32]
+        $PercentComplete,
+
+        [Parameter()]
+        [ValidateRange(0, 10)]
+        [System.UInt32]
+        $Priority
+    )
+}
+
+function Update-MGPlannerTask
+{
+    [CmdletBinding()]
+    Param(
+        [Parameter()]
+        [System.String]
+        $PlanId,
+
+        [Parameter()]
+        [System.String]
+        $Title,
+
+        [Parameter()]
+        [System.String]
+        $TaskId,
+
+        [Parameter()]
+        [System.String]
+        $Description,
+
+        [Parameter()]
+        [System.String]
+        $StartDateTime,
+
+        [Parameter()]
+        [System.String]
+        $CompletedDateTime,
+
+        [Parameter()]
+        [ValidateRange(0, 100)]
+        [System.Uint32]
+        $PercentComplete,
+
+        [Parameter()]
+        [ValidateRange(0, 10)]
+        [System.UInt32]
+        $Priority
+    )
+}
+
+function Get-MgPlannerPlanBucket
+{
+    [CmdletBinding()]
+    Param(
+        [Parameter()]
+        [System.String]
+        $PlannerPlanId
+    )
+}
+
+function New-MgPlannerBucket
+{
+    [CmdletBinding()]
+    Param(
+        [Parameter()]
+        [System.String]
+        $PlanId,
+
+        [Parameter()]
+        [System.String]
+        $Name
+    )
+}
+
+function Update-MgPlannerBucket
+{
+    [CmdletBinding()]
+    Param(
+        [Parameter()]
+        [System.String]
+        $PlanId,
+
+        [Parameter()]
+        [System.String]
+        $Name,
+
+        [Parameter()]
+        [System.String]
+        $BucketId
+    )
+}